Backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymongo==4.6.0
motor==3.3.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
Pillow==10.0.1
numpy==1.24.3
opencv-python==4.8.1.78
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
